\name{fish}
\alias{fish}
\title{Stickleback growth data
}
\description{
The data gives lengths (mm) of 53 female stickleback measured at the age of 17, 45, 80 and 115 days.}
\usage{
fish
}
\format{A dataframe with 53 observations of four variables: 

fish[,1] Length at age at 17 days

fish[,2] Length at age at 45 days

fish[,3] Length at age at 80 days

fish[,4] Length at age at 115 days
}
\references{Kuparinen A, Björklund M (2011) Theory put into practice: an R implementation of the infinite-dimensional model. Ecological Modelling (in press)
}
\seealso{\code{\link{IDM.bootCI}}}
